+++ lib/main.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:intl/date_symbol_data_local.dart';
 import 'package:provider/provider.dart';
 import 'package:stocodi_app/screens/app.dart';
-import 'package:stocodi_app/screens/invest/portfolio/data/portfolio_data.dart';
+import 'package:stocodi_app/model/portfolio/portfolio_data.dart';
 import 'package:stocodi_app/screens/sign_up/signup.dart';
 import 'package:stocodi_app/screens/sign_up/splash_screen.dart';
 import 'API/retrofit/auth_manager.dart';
