+++ lib/screens/home/widget/home_manage.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:stocodi_app/screens/invest/portfolio.dart';
-import 'package:stocodi_app/screens/invest/portfolio/data/portfolio_data.dart';
+import 'package:stocodi_app/model/portfolio/portfolio_data.dart';
 
 class HomeManage extends StatefulWidget {
   const HomeManage({super.key});
