+++ lib/screens/home/widget/home_transaction.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:stocodi_app/screens/invest/portfolio/data/portfolio_data.dart';
+import 'package:stocodi_app/model/portfolio/portfolio_data.dart';
 import 'package:stocodi_app/screens/transaction/screens/transcation_main.dart';
 
 class HomeTransaction extends StatefulWidget {
