+++ lib/screens/invest/invest_experiment.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:stocodi_app/screens/invest/portfolio/data/portfolio_data.dart';
+import 'package:stocodi_app/model/portfolio/portfolio_data.dart';
 import 'package:stocodi_app/widgets/custom_appbar.dart';
 import '../../../../theme/app_theme.dart';
 import 'Invest_experiment/data/interest_item_data_manager.dart';
@@ -46,7 +46,6 @@ class _InvestExperimentState extends State<InvestExperiment> {
               investmentItems: investmentItems,
               context: context,
             ),
-
           ],
         ),
       ),
