+++ lib/screens/transaction/screens/transcation_main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:stocodi_app/screens/invest/portfolio/data/portfolio_data.dart';
+import 'package:stocodi_app/model/portfolio/portfolio_data.dart';
 import 'package:stocodi_app/theme/transaction_theme.dart';
 import 'package:stocodi_app/widgets/custom_appbar.dart';
 import '../transaction_info.dart';
