+++ lib/test/DongGiTestPage.dart
@@ -4,9 +4,10 @@ import '../screens/lecture/classroom.dart';
 import '../screens/lecture/lecture.dart';
 import '../screens/lecture/widget/comment.dart';
 
-void main(){
+void main() {
   runApp(const DongGi());
 }
+
 class DongGi extends StatelessWidget {
   const DongGi({super.key});
 
@@ -35,25 +36,17 @@ class HomeScreen extends StatelessWidget {
               onPressed: () {
                 Navigator.push(
                   context,
-                  MaterialPageRoute(builder: (context)=> ClassRoom()),
+                  MaterialPageRoute(builder: (context) => ClassRoom()),
                 );
               },
               child: const Text('classroom'),
             ),
+            const Text('LectureActivity'),
             ElevatedButton(
               onPressed: () {
                 Navigator.push(
                   context,
-                  MaterialPageRoute(builder: (context)=> Lecture()),
-                );
-              },
-              child: const Text('LectureActivity'),
-            ),
-            ElevatedButton(
-              onPressed: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(builder: (context)=> SignUp()),
+                  MaterialPageRoute(builder: (context) => SignUp()),
                 );
               },
               child: const Text('API'),
@@ -62,12 +55,11 @@ class HomeScreen extends StatelessWidget {
               onPressed: () {
                 Navigator.push(
                   context,
-                  MaterialPageRoute(builder: (context)=> Comment()),
+                  MaterialPageRoute(builder: (context) => Comment()),
                 );
               },
               child: const Text('MyComment'),
             ),
-
           ],
         ),
       ),
